refactor(knowledge): split overlay.rs into a module directory
